feat: fix product detail page
