+++ src/routes/index.tsx
@@ -4,6 +4,7 @@ import { Layout } from "@/components/features/Layout";
 import { HomePage } from "@/pages/Home";
 import { LoginPage } from "@/pages/Login";
 import { MyAccountPage } from "@/pages/MyAccount";
+import { OrderPage } from "@/pages/Order";
 import { ProductsPage } from "@/pages/Products";
 import { ThemePage } from "@/pages/Theme";
 
@@ -36,6 +37,10 @@ const router = createBrowserRouter([
       {
         path: RouterPath.products,
         element: <ProductsPage />,
+      },
+      {
+        path: RouterPath.order,
+        element: <OrderPage />,
       },
       {
         path: RouterPath.notFound,
